# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCNU PICK 백엔드",
    "description": "프론트엔드와 분리된 백엔드 API입니다. 기능 기준은 저장소의 docs/product-scope.md이며, 프론트 연동 계약 제안은 backend/docs/API_CONTRACT.md에 있습니다.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bookmarks": {
      "get": {
        "tags": [
          "찜·캘린더"
        ],
        "summary": "본인이 찜한 공지 목록",
        "operationId": "bookmarks_api_bookmarks_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticePage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/bookmarks/{notice_id}": {
      "post": {
        "tags": [
          "찜·캘린더"
        ],
        "summary": "공지 찜하기",
        "operationId": "add_bookmark_api_bookmarks__notice_id__post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notice Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "찜·캘린더"
        ],
        "summary": "공지 찜 취소",
        "operationId": "delete_bookmark_api_bookmarks__notice_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notice Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/calendar": {
      "get": {
        "tags": [
          "찜·캘린더"
        ],
        "summary": "찜 공지의 월간 신청·행사 일정",
        "operationId": "calendar_api_calendar_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "notice_id": {
            "type": "integer",
            "title": "Notice Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "enum": [
              "application",
              "event"
            ],
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "display_start": {
            "type": "string",
            "format": "date",
            "title": "Display Start"
          },
          "display_end": {
            "type": "string",
            "format": "date",
            "title": "Display End"
          },
          "is_partial": {
            "type": "boolean",
            "title": "Is Partial"
          },
          "continues_before_month": {
            "type": "boolean",
            "title": "Continues Before Month"
          },
          "continues_after_month": {
            "type": "boolean",
            "title": "Continues After Month"
          }
        },
        "type": "object",
        "required": [
          "id",
          "notice_id",
          "title",
          "kind",
          "label",
          "start_date",
          "end_date",
          "start_time",
          "end_time",
          "display_start",
          "display_end",
          "is_partial",
          "continues_before_month",
          "continues_after_month"
        ],
        "title": "CalendarItem"
      },
      "CalendarOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "timezone": {
            "type": "string",
            "const": "Asia/Seoul",
            "title": "Timezone",
            "default": "Asia/Seoul"
          },
          "end_inclusive": {
            "type": "boolean",
            "title": "End Inclusive",
            "default": true
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CalendarItem"
            },
            "type": "array",
            "title": "Events"
          },
          "undated": {
            "items": {
              "$ref": "#/components/schemas/UndatedItem"
            },
            "type": "array",
            "title": "Undated"
          }
        },
        "type": "object",
        "required": [
          "month",
          "events",
          "undated"
        ],
        "title": "CalendarOut"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorInfo"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut"
      },
...
      "NoticeCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_code": {
            "type": "string",
            "enum": [
              "SCNU_MAIN",
              "SCNU_SW",
              "SCNU_AI"
            ],
            "title": "Source Code"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "posted_date": {
            "type": "string",
            "format": "date",
            "title": "Posted Date"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "summary_lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Summary Lines"
          },
          "deadline_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Date"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "d_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "D Day"
          },
          "deadline_label": {
            "type": "string",
            "title": "Deadline Label"
          },
          "is_closed": {
            "type": "boolean",
            "title": "Is Closed"
          },
          "is_bookmarked": {
            "type": "boolean",
            "title": "Is Bookmarked"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source_code",
          "source_name",
          "posted_date",
          "original_url",
          "category",
          "summary_lines",
          "deadline_date",
          "deadline_at",
          "d_day",
          "deadline_label",
          "is_closed",
          "is_bookmarked",
          "needs_review",
          "recommendation"
        ],
        "title": "NoticeCard"
      },
...
      "NoticePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoticeCard"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "NoticePage"
      },
...
      "RecommendationOut": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/ScorePart"
            },
            "type": "array",
            "title": "Breakdown"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          }
        },
        "type": "object",
        "required": [
          "score",
          "grade",
          "reasons",
          "breakdown",
          "policy_version"
        ],
        "title": "RecommendationOut"
      },
...
      "ScorePart": {
        "properties": {
          "criterion": {
            "type": "string",
            "title": "Criterion"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "criterion",
          "score",
          "max_score",
          "reason"
        ],
        "title": "ScorePart"
      },
...
      "UndatedItem": {
        "properties": {
          "notice_id": {
            "type": "integer",
            "title": "Notice Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "notice_id",
          "title",
          "label"
        ],
        "title": "UndatedItem"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}